load state take 2
