+++ components/card-info-overlay/index.js
@@ -23,6 +23,7 @@ export default function CardInfoOverlay({
   getCardData,
   own,
   reset,
+  load,
 }) {
   return (
     <Modal isOpen={isOpen} onClose={onClose} size={["full", "4xl"]} isCentered>
@@ -33,18 +34,22 @@ export default function CardInfoOverlay({
           <ModalCloseButton size="lg" top="11px" rounded="lg" />
 
           <ModalBody>
-            <CardInfo
-              onClose={onClose}
-              gap={gap}
-              buy={buy}
-              name={name}
-              img={img}
-              attr={attr}
-              id={id}
-              getCardData={getCardData}
-              own={own}
-              reset={reset}
-            />
+            {load ? (
+              "Loading"
+            ) : (
+              <CardInfo
+                onClose={onClose}
+                gap={gap}
+                buy={buy}
+                name={name}
+                img={img}
+                attr={attr}
+                id={id}
+                getCardData={getCardData}
+                own={own}
+                reset={reset}
+              />
+            )}
           </ModalBody>
 
           <ModalFooter>
